Move  to NotificationWithStatus from NotificationGroupWithStatus
