+++ app/javascript/mastodon/features/notifications_v2/components/notification_with_status.tsx
@@ -16,7 +16,10 @@ import Status from 'mastodon/containers/status_container';
 import { useAppSelector, useAppDispatch } from 'mastodon/store';
 
 import { DisplayedName } from './displayed_name';
-import type { LabelRenderer } from './notification_group_with_status';
+
+export type LabelRenderer = (
+  values: Record<string, React.ReactNode>,
+) => JSX.Element;
 
 export const NotificationWithStatus: React.FC<{
   type: string;
